refactor : Line Entity에 있는 Section Collection 일급 컬렉션으로 변경
