+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -4,9 +4,7 @@ import nextstep.subway.common.BaseEntity;
 import nextstep.subway.section.domain.Section;
 
 import javax.persistence.*;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Entity
 public class Line extends BaseEntity {
@@ -20,8 +18,8 @@ public class Line extends BaseEntity {
     @Column(name = "color")
     private String color;
 
-    @OneToMany(mappedBy = "line", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     protected Line() {
     }
@@ -34,26 +32,6 @@ public class Line extends BaseEntity {
     public void update(Line line) {
         this.name = line.getName();
         this.color = line.getColor();
-    }
-
-    public List<Section> getStationInOrder() {
-        List<Section> sectionsResult = new ArrayList<>();
-
-        Optional<Section> preStation = this.sections.stream()
-                .filter(st -> st.getPreStation() == null)
-                .findFirst();
-
-        while (preStation.isPresent()) {
-            final Section section = preStation.get();
-            sectionsResult.add(section);
-
-            preStation = sections.stream()
-                    .filter(st -> st.getPreStation() != null)
-                    .filter(st -> st.getPreStation().equals(section.getStation()))
-                    .findFirst();
-        }
-
-        return sectionsResult;
     }
 
     public Long getId() {
@@ -69,7 +47,10 @@ public class Line extends BaseEntity {
     }
 
     public void addSection(final Section section) {
-        this.sections.add(section);
+        this.sections.addSection(section);
     }
 
+    public List<Section> getSections() {
+        return sections.getStationInOrder();
+    }
 }
